third_party/WebKit/LayoutTests/http/tests/loading/module-script-wrong-mime-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
CONSOLE ERROR: Failed to load module script: The server responded with a non-JavaScript MIME type of "text/plain". Strict MIME type checking is enforced for module scripts per HTML spec.
main frame - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
This is a testharness.js-based test.
PASS module script with invalid mime type shouldn't have run.
Harness: the test ran to completion.
